# sql/tests: reduce variance in BenchmarkKV