This commit disables timeseries and auto SQL stats collection in the BenchmarkKV test suite. This reduces the variance in the benchmark results, allowing us to run fewer iterations to get stable results.

```
name                  old time/op    new time/op    delta
KV/Update/SQL/rows=1     219µs ± 9%     207µs ± 5%  -5.36%  (p=0.000 n=19+18)

name                  old alloc/op   new alloc/op   delta
KV/Update/SQL/rows=1    57.0kB ± 1%    56.9kB ± 0%  -0.22%  (p=0.002 n=19+20)

name                  old allocs/op  new allocs/op  delta
KV/Update/SQL/rows=1       553 ± 0%       553 ± 0%    ~     (p=0.555 n=19+17)
```

Epic: None
Release note: None